requirements.txt
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
colorama==0.4.6
colorlog==6.9.0
google-auth==2.40.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
idna==3.10
lxml==6.0.0
numpy==2.3.1
outcome==1.3.0.post0
pandas==2.3.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
